refactor: split event/block sql logic into multiple files
