+++ store/src/sql/entities/mod.rs
@@ -1,14 +1,18 @@
 mod block;
 mod event;
+mod event_block;
+mod hash;
 mod interest;
 mod query;
+mod utils;
 
 pub use block::{BlockBytes, BlockRow};
-pub use event::{
-    rebuild_car, CountRow, DeliveredEvent, EventBlockRaw, EventRaw, OrderKey, ReconHash,
-};
+pub use event::{rebuild_car, EventRaw};
+pub use event_block::EventBlockRaw;
+pub use hash::{BlockHash, ReconHash};
 pub use interest::CeramicOneInterest;
 pub use query::{BlockQuery, EventBlockQuery, EventQuery, ReconQuery, ReconType, SqlBackend};
+pub use utils::{CountRow, DeliveredEvent, OrderKey};
 
 #[derive(Debug, Clone, sqlx::FromRow)]
 pub struct FirstAndLast {
